All cent values are handled by UInt64 in Live Auction context, and a minor view update for an auction toolbar
